going over feedback and fixed
Changing the role to admin does not work.
Steps to Reproduce:

Create a user with ROLE_USER.
Connect as Admin and edit the created user: change the role to ADMIN.
Reconnect as the new user. → The user still has ROLE_USER.

A user can change their username on the edit page, but an exception occurs when clicking "Save."
Shouldn't this field be non-editable in the frontend?

UI/UX Issues:

Button styles are inconsistent on the "User Detail Edit" page. For example, the "Save" and "Cancel" buttons have different styles.
There are inconsistent cell heights in the user edit form, e.g., for the Role and Password fields.

Signed-off-by: amarhbaneen <amaralzana@gmail.com>